backend/requirements.txt
fastapi
uvicorn
pdfplumber
python-docx
pandas
chardet
python-multipart